Changed read and write operations to work in Little Endian.
Fixed ushort write operations so that half of it is no longer lost and it now works in Little Endian.

Added an project which extracts useful opcode data from http://imrannazar.com/Gameboy-Z80-Opcode-Map and saves it in an XML file.
Added an assembler as a new project which will assemble code based on the opcode data extracted from the aforementioned project.